@@ -3,12 +3,12 @@ package com.github.anidb.vo;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlElementWrapper;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlAttribute;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlElementWrapper;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 @XmlAccessorType(XmlAccessType.FIELD)
@@ -36,12 +36,43 @@ public class Anime {
 	@XmlElement(name = "episode")
 	private List<Episode> episodes = new ArrayList<>();
 
+	@XmlElementWrapper(name = "characters")
+	@XmlElement(name = "character")
+	private List<AnimeCharacter> characters = new ArrayList<>();
+
+	private String startdate;
+	private String enddate;
+
 	public void setEpisodes(List<Episode> episodes) {
 		this.episodes = episodes;
 	}
 
+	public String getStartdate() {
+		return startdate;
+	}
+
+	public void setStartdate(String startdate) {
+		this.startdate = startdate;
+	}
+
+	public String getEnddate() {
+		return enddate;
+	}
+
+	public void setEnddate(String enddate) {
+		this.enddate = enddate;
+	}
+
 	public String getPicture() {
 		return picture;
+	}
+
+	public List<AnimeCharacter> getCharacters() {
+		return characters;
+	}
+
+	public void setCharacters(List<AnimeCharacter> characters) {
+		this.characters = characters;
 	}
 
 	public void setPicture(String picture) {
