@@ -16,19 +16,22 @@ import java.net.URL;
 import java.net.URLConnection;
 import java.security.SecureRandom;
 import java.security.cert.X509Certificate;
-import java.util.logging.Logger;
 import java.util.zip.GZIPInputStream;
 
 import javax.net.ssl.HttpsURLConnection;
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManager;
 import javax.net.ssl.X509TrustManager;
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBException;
-import javax.xml.bind.UnmarshalException;
-import javax.xml.bind.Unmarshaller;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import com.github.anidb.vo.Anime;
+
+import jakarta.xml.bind.JAXBContext;
+import jakarta.xml.bind.JAXBException;
+import jakarta.xml.bind.UnmarshalException;
+import jakarta.xml.bind.Unmarshaller;
 
 /**
  * See: http://wiki.anidb.net/w/HTTP_API_Definition
@@ -49,12 +52,15 @@ public class AnidDB {
 	 */
 	private int timeOut = 5000;
 
-	private static final String CONNECTION_URL = BASE_URL + "client=" + CLIENT_NAME + "&clientver=" + VERSION + "&protover=1";
+	private static final String CONNECTION_URL = BASE_URL + "client=" + CLIENT_NAME + "&clientver=" + VERSION
+			+ "&protover=1";
 
 	private Proxy proxy;
 
 	private URLConnection anidbConnection;
-	private boolean disableSSLTrustmanager = true;
+	private boolean disableSSLTrustmanager = false;
+	private static final Logger LOG = LoggerFactory.getLogger(AnidDB.class);
+	
 
 	/**
 	 * 
@@ -74,13 +80,16 @@ public class AnidDB {
 	private void disableSSLTrustManager() {
 		TrustManager[] trustAllCerts = new TrustManager[] { new X509TrustManager() {
 
+			@Override
 			public X509Certificate[] getAcceptedIssuers() {
 				return null;
 			}
 
+			@Override
 			public void checkClientTrusted(X509Certificate[] certs, String authType) {
 			}
 
+			@Override
 			public void checkServerTrusted(X509Certificate[] certs, String authType) {
 			}
 		} };
@@ -95,8 +104,7 @@ public class AnidDB {
 
 	/**
 	 * 
-	 * @param timeOut
-	 *            Timeout default 5000
+	 * @param timeOut Timeout default 5000
 	 */
 	public void setTimeOut(int timeOut) {
 		this.timeOut = timeOut;
@@ -118,13 +126,14 @@ public class AnidDB {
 		}
 		try {
 			URL url = new URL(urlSb.toString());
-			Logger.getLogger(getClass().getName()).info("Connect:" + urlSb);
+			LOG.info("Connect:" + urlSb);
 			if (proxy != null) {
-				anidbConnection = (HttpURLConnection) url.openConnection(proxy);
+				anidbConnection = url.openConnection(proxy);
 			} else {
-				anidbConnection = (HttpURLConnection) url.openConnection();
+				anidbConnection = url.openConnection();
 			}
-			anidbConnection.setRequestProperty("User-Agent", "Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10.4; en-US; rv:1.9.2.2) Gecko/20100316 Firefox/3.6.2");
+			anidbConnection.setRequestProperty("User-Agent",
+					"Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10.4; en-US; rv:1.9.2.2) Gecko/20100316 Firefox/3.6.2");
 			anidbConnection.setRequestProperty("Referer", "http://anidb.net/perl-bin/animedb.pl?show=main");
 			anidbConnection.setConnectTimeout(timeOut);
 			anidbConnection.setReadTimeout(timeOut);
@@ -154,8 +163,37 @@ public class AnidDB {
 			throw new IllegalArgumentException("anidbIdparameter is wrong");
 		}
 		try {
+			if (writeFile) {
+				File cache = new File(anidbId + ".xml");
+				if (cache.exists()) {
+					try (FileInputStream readSource = new FileInputStream(cache);
+							InputStreamReader isr = new InputStreamReader(readSource)) {
+						return readDataAsString(isr);
+					} catch (IOException e) {
+						throw new RuntimeException(e);
+					}
+				}
+			}
 			connect(new URLParameter("request", "anime"), new URLParameter("aid", Integer.toString(anidbId)));
 			String xmlData = readDataStringFromConnection();
+			if (writeFile) {
+				FileOutputStream fos = null;
+				try {
+					fos = new FileOutputStream(new File(anidbId + ".xml"));
+
+					fos.write(xmlData.getBytes());
+				} catch (IOException e1) {
+					LOG.error(e1.getLocalizedMessage(),e1);
+				} finally {
+					if (fos != null) {
+						try {
+							fos.close();
+						} catch (IOException e1) {
+							LOG.error(e1.getLocalizedMessage(),e1);
+						}
+					}
+				}
+			}
 			return xmlData;
 		} finally {
 			cleanup();
@@ -165,8 +203,7 @@ public class AnidDB {
 	/**
 	 * 
 	 * @return Anime from httpapi.xml
-	 * @throws IOException
-	 *             Loadingerror
+	 * @throws IOException Loadingerror
 	 */
 	public Anime test() throws IOException {
 		try {
@@ -185,8 +222,7 @@ public class AnidDB {
 	/**
 	 * Loading the animedata from anidb.net
 	 * 
-	 * @param anidbId
-	 *            ID
+	 * @param anidbId ID
 	 * @return Anime
 	 */
 	public Anime loadAnime(int anidbId) {
@@ -197,40 +233,33 @@ public class AnidDB {
 				try {
 					readSource = new FileInputStream(cache);
 				} catch (FileNotFoundException e) {
-					throw new RuntimeException(e); 
+					throw new RuntimeException(e);
 				}
 			}
 		}
 		if (readSource == null) {
 			String xml = loadAnimeXML(anidbId);
-			if (writeFile) {
-				FileOutputStream fos = null;
-				try {
-					fos = new FileOutputStream(new File(anidbId + ".xml"));
 
-					fos.write(xml.getBytes());
-				} catch (IOException e1) {
-					// TODO Auto-generated catch block
-					e1.printStackTrace();
-				} finally {
-					if (fos != null) {
-						try {
-							fos.close();
-						} catch (IOException e1) {
-							// TODO Auto-generated catch block
-							e1.printStackTrace();
-						}
-					}
-				}
-			}
 			try {
 				readSource = new ByteArrayInputStream(xml.getBytes("UTF-8"));
 			} catch (UnsupportedEncodingException e) {
-				// TODO Auto-generated catch block
-				e.printStackTrace();
+				LOG.error(e.getLocalizedMessage(),e);
 			}
 		}
-		
+
+		return convertXMLToAnime(readSource);
+	}
+
+	public Anime convertXMLToAnime(String xml) {
+		try (ByteArrayInputStream readSource = new ByteArrayInputStream(xml.getBytes("UTF-8"))) {
+			return convertXMLToAnime(readSource);
+		} catch (IOException e) {
+			LOG.error(e.getLocalizedMessage(),e);
+			throw new RuntimeException(e);
+		}
+	}
+
+	public Anime convertXMLToAnime(InputStream readSource) {
 		try {
 
 			JAXBContext context = JAXBContext.newInstance(Anime.class);
@@ -239,13 +268,13 @@ public class AnidDB {
 			return param;
 		} catch (UnmarshalException e) {
 
-			Logger.getLogger(getClass().getName()).info("Could not read fron anidb:" + e.getLocalizedMessage());
+			LOG.info("Could not read fron anidb:" + e.getLocalizedMessage());
 			e.printStackTrace();
 
 			return null;
 		} catch (JAXBException e) {
 			throw new RuntimeException(e);
-		}  
+		}
 	}
 
 	/**
@@ -267,35 +296,39 @@ public class AnidDB {
 	}
 
 	private String readDataFromConnection() {
-		InputStreamReader is = null;
-		try {
-			if (anidbConnection == null) {
-				throw new IllegalStateException("No active connection found");
-			}
-			if ("gzip".equalsIgnoreCase(anidbConnection.getHeaderField("Content-Encoding"))) {
-				is = new InputStreamReader(new GZIPInputStream(anidbConnection.getInputStream()), "UTF-8");
-			} else {
-				is = new InputStreamReader(anidbConnection.getInputStream(), "UTF-8");
-			}
-			int read = 0;
-			char[] data = new char[1024];
-			StringBuffer baos = new StringBuffer();
-			while ((read = is.read(data)) != -1) {
-				baos.append(data, 0, read);
-			}
-			return baos.toString();
-		} catch (IOException e) {
-			throw new RuntimeException(e);
-		} finally {
-			if (is != null) {
-				try {
-					is.close();
-				} catch (IOException e) {
 
+		if (anidbConnection == null) {
+			throw new IllegalStateException("No active connection found");
+		}
+		try (InputStream inputStream = anidbConnection.getInputStream()) {
+
+			if ("gzip".equalsIgnoreCase(anidbConnection.getHeaderField("Content-Encoding"))) {
+				try (GZIPInputStream gzip = new GZIPInputStream(inputStream);
+						InputStreamReader isr = new InputStreamReader(gzip, "UTF-8")) {
+
+					return readDataAsString(isr);
+				}
+			} else {
+				try (InputStreamReader isr = new InputStreamReader(inputStream, "UTF-8")) {
+
+					return readDataAsString(isr);
 				}
 			}
+
+		} catch (IOException e) {
+			throw new RuntimeException(e);
 		}
 
+	}
+
+	public String readDataAsString(InputStreamReader is) throws IOException {
+		int read = 0;
+		char[] data = new char[1024];
+		StringBuffer baos = new StringBuffer();
+		while ((read = is.read(data)) != -1) {
+			baos.append(data, 0, read);
+		}
+		return baos.toString();
 	}
 
 	/**
